mangrove.js: Improve CLI useability
The CLI is now compiled and can be run without ts-node.
This also means that CI will complain if we forget to update it.

As the CLI includes package.json, we've had to move tsconfig.json to src/ and introduce a simpler one in the root of mangrove.js. Otherwise, the compiled files would be placed in dist/nodejs/src.
See the following link for more information: https://stackoverflow.com/a/61467483/16614325