feat(Greedy Taker bot): Use Toad Scheduler instead of `sleep` and rely on nodejs shutdown
According to the Nodejs documentation, calling `process.exit()` is bad practice: https://nodejs.org/api/process.html#processexitcode
One risk is that any logs written before exiting may not be written.

The process shutdown logic has therefore been changed to stop all takers and thereby remove all work from the event loop which should cause Nodejs to shut down.

As part of this, the `OfferTaker` class has been changed to use ToadScheduler instead of an infinite loop with a `sleep` call; This is easier to stop immediately and more flexible (and we use it in other bots).