feat: Set MgvOracle mutator to gasbot on deploy
Previously, the MgvOracle mutator was set to `chief` on deploy and then changed to gasbot when Mangrove was configured to use the oracle.
Now, the mutator is immediately set to gasbot. Also, the script to configure Mangrove to use an oracle has been simplified and now supports any oracle, not just MgvOracle.