feat: Make scripts consistent
The following conventions have been applied to all *.s.sol scripts:

1. Env vars are read in the `run` function, not the `innerRun` function.
This ensures that env vars are only read by the script the user invoked and will not accidentally be picked up (and possibly misinterpreted) by another script.

2. Env vars specifying contracts can always contain either a contract name or an address.
This functionality (provided by the envAddressOrName function) was used in some places, but not all, resulting in an inconsistent user experience.
Contract names are resolved via the `fork.get` function, ie the ToyENS registry.

3. Contract names are resolved in the `run` function, not the `innerRun` function.
This helps ensure that contract addresses specified by the user when calling the outermost script are not ignored by another script (which could happen before because scripts relied on the address provided by fork.get for a hardcoded name).

4. Contract addresses can always be specified by the user via env vars.
Instead of always of relying on all addresses being registered in ToyENS and looking these up via hardcoded names, all scripts now allow contract addresses/names to be specified in env vars.

5. Specific contract/interface types are preferred for `innerRun` parameters instead of `address`.
This reduces the risk of passing a wrong a address from another script.

6. `outputDeployment` is called in the `run` function, not in `innerRun`
This ensures the outermost script is in control of when the deployment is output.