docs(license): Replace UNLICENSED with Unlicense
'UNLICENSED' is an invalid SPDX identifier with ambiguous meaning: Some interpret it as public domain others as the opposite, i.e. proprietary.

Our intent has been for files tagged with UNLICENSED to be public domain. The Unlicense license captures that well, so we've put all files previously tagged with UNLICENSED under the Unlicense license.