fix(scripts): Workaround for forge bug in KandelSeederDeployer
forge has a bug where it fails to increment the broadcaster nonce when deploying a contract which itself deploys a contract.
This is bug is triggered when deploying AaveKandelSeeder which deploys a router contract.

A workaround has been implemented which forces the nonce to be incremented if forge has not done it.