fix: correct posthook values for sor.wants and sor.gives
The previously introduced optimization not restoring `sor.wants`/`gives`
values before a posthook when an offer has been partially filled does
not work when the partial fill has failed.

In that case, execution continues (despite the existence of a partial
fill), and so the partial-fill specific values should be restored
anyway.

Added a regression test in MakerPosthook.