refactor: Simplify offer price match flow
Previously, `execute` checked that the offer to be executed had a sufficiently low price. If this wasn't the case, "mgv/notExecuted" was returned and used to tell the next recursive call to `internalMarketOrder` that the order was finished.

This flow was somewhat convoluted: The decision on which offers to execute was distributed between multiple functions, despite only the outermost of those actually needed it; `execute` had two if-stmts just to make sure it **did not** handle an offer that shouldn't be executed.

This commit instead moves the check out one level to `internalMarketOrder` and `clean`. This simplifies the code of all involved functions and reduces gas usage.
The tradeoff is that the price check code (a single function call) is duplicated and that the caller of `execute` (which is internal) must ensure that the taker is not matched with an offer that is too expensive. However, this is seems acceptable as the duplicated code is a trivial function call and `execute` is `internal`.