feat: Post offers at nearest, higher tick for unaligned logPrices
The maker API admits (log)Prices that do not correspond exactly to a tick. Frontends and other intermediaries should ensure offers aren't posted at such prices, but we cannot ensure this. So either the maker API should revert if given such an unaligned logPrice or the offer should be posted at either the nearest higher or lower tick.

We think that some integrations will be easier if the API is permissive and allows unaligned logPrices, so we opt for choosing one of the nearest ticks.

Previously, the nearest, lower tick was chosen when adding the offer to the book. This meant that makers would sell at a lower price than what they specified which seems wrong.

This commit therefore changes the maker API such that offers are posted at the nearest, higher tick when given an unaligned logPrice.

Also, this commit optimized the conversion of offer logPrices to ticks, by exploiting that offers are always stored with a tick-aligned logPrice.