test(ratio): Fix use of vm.expectRevert in TickAndBin tests
vm.expertRevert has surprising behavior when used before non-external calls: Anything after the call will not execute.

See this GitHub issue and comment for more details: https://github.com/foundry-rs/foundry/issues/3437#issuecomment-1336178807

Expecting reverts for non-external calls is therefore only safe if the call is last in the test.

This commit fixes two tests in TickAndBinTest which assumed that multiple reverts could be expected within a single test. The tests are split into one test per expected revert.

The commit also fixes an error in one of the test cases which this change uncovered:
The test assumed that it was possible to pass a too small `exp` to `TickLib.tickFromNormalizedRatio`. However, MAX_RATIO_EXP is zero and the exp parameter is uint,
so it is not possible pass an exp value that is smaller than MAX_RATIO_EXP, since uint(MAX_RATIO_EXP - 1) = uint(-1) = type(uint)max.